requirements.txt
transformers
sentence-transformers
faiss-cpu
requests
spacy
torch
datasets
torch